requirements.txt

matplotlib-inline==0.1.7
numpy==1.24.2
pandas==2.2.2
seaborn==0.12.2
streamlit